requirements.txt
appdirs==1.4.4
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.2
exceptiongroup==1.1.0
frozendict==2.3.5
greenlet==2.0.2
html5lib==1.1
idna==3.4
iniconfig==2.0.0
lxml==4.9.2
multitasking==0.0.11
mysql-connector-python==8.0.32
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pluggy==1.0.0
protobuf==3.20.3
pycparser==2.21
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.5.post1
TA-Lib @ file:///E:\Python\API\CryptoCurrency\botTrader\Crypto_MySQL_2\Tools\TA_Lib-0.4.24-cp310-cp310-win_amd64.whl
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
webencodings==0.5.1
yfinance==0.2.12
